# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 3 Batti Betting Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /user/change-mpin:
    post:
      operationId: changeMpin
      tags: [user]
      summary: Change MPIN
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeMpinBody"
      responses:
        "200":
          description: MPIN changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # WALLET
  /user/wallet:
    get:
      operationId: getWallet
      tags: [wallet]
      summary: Get wallet balance
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Wallet info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /user/transactions:
    get:
      operationId: getTransactions
      tags: [wallet]
      summary: Get transaction history
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [all, credit, debit, pending]
          required: false
        - name: page
          in: query
          schema:
            type: integer
          required: false
        - name: limit
          in: query
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
  /user/bank-details:
    get:
      operationId: getBankDetails
      tags: [user]
      summary: Get bank/UPI details
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bank details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankDetails"
    put:
      operationId: updateBankDetails
      tags: [user]
      summary: Update bank/UPI details
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBankDetailsBody"
      responses:
        "200":
          description: Bank details updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankDetails"

  # GAME
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
      required: [message, success]

    # AUTH
...
    ChangeMpinBody:
      type: object
      properties:
        oldMpin:
          type: string
        newMpin:
          type: string
      required: [oldMpin, newMpin]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: string
        email:
          type: string
        referralCode:
          type: string
        kycStatus:
          type: string
          enum: [pending, verified]
        isBlocked:
          type: boolean
        walletBalance:
          type: number
        createdAt:
          type: string
      required: [id, name, mobile, referralCode, kycStatus, isBlocked, walletBalance, createdAt]
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
    Wallet:
      type: object
      properties:
        balance:
          type: number
        totalDeposited:
          type: number
        totalWithdrawn:
          type: number
        totalWon:
          type: number
        totalLost:
          type: number
      required: [balance, totalDeposited, totalWithdrawn, totalWon, totalLost]
    BankDetails:
      type: object
      properties:
        accountNumber:
          type: string
        ifsc:
          type: string
        accountHolderName:
          type: string
        bankName:
          type: string
        upiId:
          type: string
    UpdateBankDetailsBody:
      type: object
      properties:
        accountNumber:
          type: string
        ifsc:
          type: string
        accountHolderName:
          type: string
        bankName:
          type: string
        upiId:
          type: string

    # TRANSACTIONS
    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, bet, win, referral, bonus]
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected, processing, completed]
        referenceId:
          type: string
        screenshotUrl:
          type: string
        reason:
          type: string
        createdAt:
          type: string
        userName:
          type: string
        userMobile:
          type: string
      required: [id, userId, type, amount, status, createdAt]
    TransactionList:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [transactions, total, page, limit]

    # GAME
...